+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Fixed to ignore zombie or not registered replicas in `internal/admin/stats.Repository`. (#VARLOG-756/#674)
 
 
 ## [0.1.10] - 2022-06-09